8.Evaluation with TruLens/requirements.txt
llama-index
openai
chainlit
pypdf
llmsherpa
pinecone-client
llama-hub
unstructured
trulens_eval
tiktoken